requirements.txt
django==1.4.20
djangorestframework==3.1.1
MySQL-python
South
celery
redis
django-templatetag-handlebars
pytest
selenium
requests